+++ makefiles/ps.c
@@ -1,13 +1,4 @@
-#define MAX_LEN 20
-struct ps_info *trav_dir(char dir[]);
-int read_info(char d_name[],struct ps_info *p1);
-void print_ps(struct ps_info *head);
-int is_num(char *);
-void uid_to_name(uid_t uid, struct ps_info *p1);
-
-int bufsize = 1024;
-char *buffer_pwd;
-
+#include "ps.h"
 void record_pwd(){
     buffer_pwd = (char *)malloc(sizeof(char)*bufsize);
     if (!buffer_pwd){
@@ -28,17 +19,6 @@ void cd_pwd(){
     chdir(buffer_pwd);
     free(buffer_pwd);
 }
-
-typedef struct ps_info
-{
-    char pname[MAX_LEN];
-    char user[MAX_LEN];
-    int  pid;
-    int  ppid;
-    char state;
-    struct ps_info *next;
-}mps;
-
 
 int fun_ps(char** char_list)
 {
